Project 1 (Nick Scribner)
I know that this is very late, but on time. But as much as I could do and with help from my partner. I could not fully complete this code. Some things just didn't want to work and was honestly difficult to figure out on my own. Let's hope the next one isn't as ridiculous.